// packages/cantonjs-splice-validator/src/generated/scanProxy.types.ts
export interface operations {
// ...
    lookupFeaturedAppRight: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                provider_party_id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LookupFeaturedAppRightResponse"];
                };
            };
        };
    };
// ...
    getAmuletRules: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetAmuletRulesProxyResponse"];
                };
            };
        };
    };
// ...
    getAnsRules: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["GetAnsRulesRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetAnsRulesResponse"];
                };
            };
        };
    };
    lookupTransferPreapprovalByParty: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                party: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LookupTransferPreapprovalByPartyResponse"];
                };
            };
            404: components["responses"]["404"];
        };
    };
    lookupTransferCommandCounterByParty: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                party: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LookupTransferCommandCounterByPartyResponse"];
                };
            };
            /** @description No TransferCommandCounter exists for this party. This means the nonce that should be used is 0. */
            404: components["responses"]["404"];
        };
    };
    lookupTransferCommandStatus: {
        parameters: {
            query: {
                sender: string;
                nonce: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LookupTransferCommandStatusResponse"];
                };
            };
            /** @description No TransferCommand exists with this contract id within the last 24h */
            404: components["responses"]["404"];
        };
    };
// ...
}